adjusted chain total value calc formatting to return integer

+++ ChainTotalValueCalculation/CalcTools/getPrices.js
@@ -72,8 +72,7 @@ const getPrices = async (networkId, array) => {
   }
 
   //Returns final total, rounded to 0 places and formatted
-  grandTotal = parseFloat(grandTotal).toFixed(0);
-  // grandTotal = new Intl.NumberFormat({ style: "currency", currency: "USD" }).format(grandTotal);
+  grandTotal = parseInt(grandTotal.toFixed(0));
 
   return grandTotal;
 };
